Refactoring following some of the PR comments.  Reduced the load on the routes.service  Moved some indoor responsibility into the indoor-functions service  Abstracted the shortest path algorithm relatively  Moved out some array utils into a shared class
