+++ src/app/core/models/indoor-coordinates.ts
@@ -7,3 +7,11 @@ export interface StoredCoordinates {
     lng: string;
     fN: number;
 }
+
+export interface AdjacencyMatrix {
+    [coordinateName: string]: Edge;
+}
+
+export interface Edge {
+    [coordinateName: string]: number;
+}
