+++ src/app/shared/arrays-util.ts
@@ -1,0 +1,16 @@
+export class ArraysUtil {
+    /**
+     * Helper function to compare if two string arrays are equal.
+     */
+    static compare(array1: string[], array2: string[]) {
+        if (array1.length !== array2.length) {
+            return false;
+        }
+        for (let i = 0; i < array1.length; i++) {
+            if (array1[i] !== array2[i]) {
+                return false;
+            }
+        }
+        return true;
+    }
+}
